requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
python-dotenv>=1.0.0
requests>=2.28.2
gunicorn>=20.1.0
supabase>=1.0.3  # Add this line for Supabase support
google-generativeai>=0.3.0  # Google's official Generative AI package